pyproject.toml
[tool.poetry]
name = "chat-app"
version = "0.1.0"
description = "Application for chat with other users in real time with socket.io"
authors = ["jsiona <jimbosiona@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
dependency-injector = "^4.41.0"
aiopg = "^1.4.0"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
python-dotenv = "^0.21.1"
pydantic = "^1.10.4"
yoyo-migrations = "^8.2.0"
python-socketio = "^5.7.2"
websockets = "^10.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"
autoflake = "^2.0.0"
isort = "^5.12.0"
docformatter = "^1.5.1"
add-trailing-comma = "^2.4.0"
safety = "^2.3.5"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
